requirements.txt
fastapi
authlib
itsdangerous
ffmpeg-python
httpx
neo4j
openai
psycopg2-binary
pyaudio
pydub
pydantic
python-dotenv
redis
rq
simpleaudio
sqlalchemy
weasyprint
markdown
jwt
literalai
python-multipart
uvicorn